+++ Editor/d4rkAvatarOptimizer.cs
@@ -4726,6 +4726,9 @@ public class d4rkAvatarOptimizer : MonoBehaviour, VRC.SDKBase.IEditorOnly
                     }
                 }
             }
+
+            // clear property block to make editor testing match game behavior
+            meshRenderer.SetPropertyBlock(null);
         }
 
         foreach (var mat in optimizedMaterials.Select(o => o.target))
@@ -5253,8 +5256,6 @@ public class d4rkAvatarOptimizer : MonoBehaviour, VRC.SDKBase.IEditorOnly
             }
 
             meshRenderer.sharedMaterials = optimizedMaterials;
-            // clear property block to make editor testing match game behavior
-            meshRenderer.SetPropertyBlock(null);
 
             foreach (var ps in GetParticleSystemsUsingRenderer(meshRenderer))
             {